fix: Shadowing 점수 반환 수정(5점 만점)

+++ BackEnd/SpringBootServer/src/main/java/com/ssafy/backend/domain/news/dto/ShadowingResponse.java
@@ -1,4 +1,4 @@
 package com.ssafy.backend.domain.news.dto;
 
-public record ShadowingResponse(double similarity) {
+public record ShadowingResponse(int similarity) {
 }
